Remove import of torch in conftest.py
This fixture that was being created is only used in one place.
We need to make sure conftest can run even without backends installed.
